cyber_gym/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-cyber-gym"
version = "0.1.0"
description = "CyberGym: An uncertainty-aware tool-use environment for training agentic AI"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "flask>=3.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "cyber_gym.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["cyber_gym", "cyber_gym.server"]
package-dir = { "cyber_gym" = ".", "cyber_gym.server" = "server" }
